QuantumPrisonersDilemmaModel.QPDM — Type

QPDM{T<:Real} <: AbstractQPDM

A model object for the Quantum Prisoner's Dilemma Model. The QPDM has four basis states:

opponent defects and you defect

opponent defects and you cooperate

opponent cooperates and you defect

opponent cooperates and you cooperate

The bases are orthonormal and in standard form. The model assumes three conditions:

Player 1 (you) is told that player 2 defected

Player 1 (you) is told that player 2 cooperated

Player 1 (you) is not informed of player's action

Model inputs and outputs are assumed to be in the order above.

Fields

μd: utility for defecting

μc: utility for cooperating

γ: entanglement parameter for beliefs and actions

Example

using QuantumPrisonersDilemmaModel
model = QPDM(;μd=.51, γ=2.09)

References

Pothos, E. M., & Busemeyer, J. R. (2009). A quantum probability explanation for violations of ‘rational’decision theory. Proceedings of the Royal Society B: Biological Sciences, 276(1665), 2171-2178.

Base.rand — Method

rand(dist::AbstractQPDM, n::Int; t = π / 2)

Generates simulated data for the following conditions:

Player 2 is told that player 1 defected

Player 2 is told that player 1 cooperated

Player 2 is not informed of player 1's action

Arguments

dist::AbstractQPDM

n: the number of trials per condition

Keywords

t = π / 2: time of decision

Example

using QuantumPrisonersDilemmaModel 
model = QPDM(;μd=.51, γ=2.09)
data = rand(model, 100)

Distributions.logpdf — Method

logpdf(dist::AbstractQPDM, n::Int, n_d::Vector{Int}; t = π / 2)

Returns the joint log density given data for the following conditions:

Player 2 is told that player 1 defected

Player 2 is told that player 1 cooperated

Player 2 is not informed of player 1's action

Arguments

dist::AbstractQPDM

n: the number of trials per condition

n_d: the number of defections in each condition

Keywords

t = π / 2: time of decision

Example

using QuantumPrisonersDilemmaModel 
model = QPDM(;μd=.51, γ=2.09)
n_trials = 100
data = rand(model, n_trials)
logpdf(model, n_trials, data)

Distributions.pdf — Method

pdf(dist::AbstractQPDM, n::Int, n_d::Vector{Int}; t = π / 2)

Returns the joint probability density given data for the following conditions:

Player 2 is told that player 1 defected

Player 2 is told that player 1 cooperated

Player 2 is not informed of player 1's action

Arguments

dist::AbstractQPDM

n: the number of trials per condition

n_d: the number of defections in each condition

Keywords

t = π / 2: time of decision

QuantumPrisonersDilemmaModel.predict — Method

predict(dist::AbstractQPDM; t = π / 2)

Returns predicted response probability for the following conditions:

Player 2 is told that player 1 defected

Player 2 is told that player 1 cooperated

Player 2 is not informed of player 1's action

Arguments

dist::AbstractQPDM

Keywords

t = π / 2: time of decision

Example

using QuantumPrisonersDilemmaModel 
model = QPDM(;μd=.51, γ=2.09)
predict(model)